fix: parse forked bazel versions (#6)
## Problem

Was doing some more work with bazel and was curious whether this lsp had
updates. I was working with a forked version of bazel and hard coded the
versions array to include mine. This seemed to cause the build to fail
after updating.

The root cause was that version included a `/` to signal to `bazelisk`
that it should pull the bazel version from a github repo. The caused the
build to fail due to a parent directory of the flags cache not existing.

## Solution

I added a check to ensure that the parent directory of the flags cache
exists before attempting to create/write to it. Also, added the error
message in the build step for easier future debugging.